Use half precision for NeuralNetwork weights (50% memory savings)

+++ source/EngineGpuKernels/Entities.cuh
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cuda_fp16.h>
+
 #include <EngineInterface/CudaSettings.h>
 #include <EngineInterface/CellTypeConstants.h>
 #include <EngineInterface/EngineConstants.h>
@@ -48,9 +50,12 @@ struct ObjectConnection
     float angleFromPrevious;
 };
 
+// NeuralNetwork weights stored as half precision (FP16) for memory efficiency
+// and optimal tensor core usage. Saves 50% memory compared to float storage.
+// With MAX_CHANNELS=16: weights = 256 half = 512 bytes (was 1024 bytes with float)
 struct NeuralNetwork
 {
-    float weights[MAX_CHANNELS * MAX_CHANNELS];
+    half weights[MAX_CHANNELS * MAX_CHANNELS];  // FP16 for tensor cores
     float biases[MAX_CHANNELS];
     ActivationFunction activationFunctions[MAX_CHANNELS];
     float connectionWeights[MAX_OBJECT_CONNECTIONS];
